requirements.txt
selenium==4.13.0
requests==2.31.0
python-dotenv~=1.0.0
pandas==2.1.1


# Segment Analytics Requirements
aiohttp==3.7.4.post0
appdirs==1.4.4
astroid==1.6.6
async-timeout==3.0.1
attrs==23.1.0
backoff==1.10.0
bleach==6.0.0
botocore==1.29.40
Cerberus==1.3.5
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
colorama==0.4.4
dephell==0.8.3
dephell-archive==0.1.7
dephell-argparse==0.1.3
dephell-changelogs==0.0.1
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-setuptools==0.2.5
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.18
dephell-versioning==0.1.2
docutils==0.20.1
entrypoints==0.3
flake8==3.7.9
git-remote-codecommit==1.16
idna==3.4
importlib-metadata==4.11.2
isort==5.9.3
Jinja2==3.0.1
jmespath==1.0.1
keyring==23.5.0
lazy-object-proxy==1.9.0
m2r==0.2.1
MarkupSafe==2.1.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
monotonic==1.6
packaging==21.0
pbr==5.11.1
pexpect==4.8.0
pkginfo==1.9.6
protobuf==4.24.2
ptyprocess==0.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.16.1
pylint==1.9.3
pyparsing==3.1.1
python-dateutil==2.8.2
readme-renderer==32.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
shellingham==1.4.0
six==1.15.0
termcolor==1.1.0
tomlkit==0.7.2
tqdm==4.63.0
twine==3.8.0
typing-extensions==3.10.0.2
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.12.1
yarl==1.9.2
yaspin==2.1.0
zipp==3.7.0